Deserialize Option and newtypes transparently
When deserializing an nvlist, if the struct calls for an Option or a
newtype, expect simply the wrapped value in the nvlist (or absent from
the nvlist in the case of an Option).  This allows us to deserialize a
DATA_TYPE_UINT64 nvpair as a `struct NewType(u64)` or an `Option<u64>`.